requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.41.0
typing-extensions>=4.9.0
urllib3>=2.1.0
openai>=1.12.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
autogen[crawl4ai]>=0.2.0
pydantic>=2.0.0
nest-asyncio>=1.5.8
networkx>=3.0
matplotlib>=3.7.0
tqdm>=4.65.0 